Remove "etab" from list of delivered files
This change removes the "etab" file from the list of files delivered by
the "nfs-kernel-server" package.

Previously the "nfs-kernel-server" package had the "etab" file delivered
as part of the package. Thus, when upgrading the package, the existing
"etab" file would get replaced with a new/empty "etab" file from the new
package.

This file is dynamically modified during runtime, to contain the list of
currently exported filesystems. Thus, when it's replaced on upgrade with
an empty file, this results in all exports being unexported; i.e. mountd
notices that the file is empty, interprets that to mean no filesystems
should be currently exported, and then unexports all previously exported
filesystems.

This is problematic, since the act of unexporting the filesystems can
result in client errors, if those exports are being actively used at the
time of the upgrade/unexport.